+++ resources/js/pages/social/components/groups/page-components/AdminActions.vue
@@ -1,31 +1,67 @@
 <template>
     <div>
         <div class="content-block">
-            <p><b>{{ $t('your-rank')}}</b>: 
-                <FaIcon :icon="group.rank == 'admin' ? 'angles-down' : 'angle-down'" />
-                {{group.rank}}
-            </p>
-            <p><b>{{ $t('joined')}}</b>: {{parseDateTime(group.joined)}} ({{daysSince(group.joined)}})</p>
+            <h4>{{ $t('manage-group')}}</h4>
+            <form @submit.prevent>
+                <label for="edit-name-comp">{{$t('name')}}</label>
+                <Editable 
+                    id="edit-name-comp" 
+                    :key="group.name" 
+                    class="ml-1 mb-2"
+                    :item="group.name" 
+                    :index="1" 
+                    :name="'name'" 
+                    @save="save" />
+                
+                <label for="edit-description-comp">{{$t('description')}}</label>
+                <Editable 
+                    id="edit-description-comp"
+                    :key="group.description"
+                    class="ml-1 mb-2"
+                    :item="group.description" 
+                    :index="2" 
+                    name="description" 
+                    type="textarea"
+                    :rows=3
+                    @save="save" />
+
+                <button class="m-1" @click="togglePublic">
+                    {{group.is_public ? 'Set to private' : 'Make public' }}
+                </button>
+                <button v-if="group.is_public" class="m-1" @click="toggleRequireApplication()">
+                    {{group.require_application ? 'Do not require application' : 'Require application' }}
+                </button>
+                <!-- TODO turn this into a 'turn off button' -->
+            </form>
+        </div>
+        <div v-if="group.require_application" class="content-block">
+            <h4>{{ $t('manage-group-applications') }}</h4>
+            <Loading v-if="loading" />
+            <template v-for="application in applications" :key="application.id">
+                <p>
+                    {{`${application.username}, applied on: ${application.applied_at}`}}
+                    <button class="m-1" @click="acceptApplication(application.id)">
+                        {{ $t('accept-group-application') }}
+                    </button>
+                    <button class="m-1" @click="rejectApplication(application.id)">
+                        {{ $t('reject-group-application') }}
+                    </button>
+                    <button class="m-1" @click="suspendApplication(application.id)">
+                        {{ $t('suspend-group-application') }}
+                    </button>
+                </p>
+            </template>
+            <div v-if="!loading && applications && applications.length < 1">
+                {{$t('no-applications')}}
+            </div>
         </div>
         <div class="d-flex m-2">
             <div v-if="group.rank == 'admin'">
                 <button type="button" class="m-1 box-shadow" @click="deleteGroup()">{{ $t('delete-group') }}</button>
-                <button type="button" class="m-1 box-shadow" @click="manageGroup()">{{ $t('manage-group') }}</button>
-                <button type="button" class="m-1 box-shadow" @click="manageApplications()">
-                    {{ $t('manage-group-applications') }}
-                </button>
                 <button type="button" class="m-1 box-shadow" @click="inviteUsers()">{{ $t('invite-users') }}</button>
                 <button type="button" class="m-1 box-shadow" @click="showBlocklist()">{{$t('blocklist')}}</button>
             </div>
         </div>
-        <Modal class="xl" :show="showManageGroupModal" 
-               :title="group.name" @close="closeManageGroup">
-            <ManageGroupModal :group="group" />
-        </Modal>
-        <Modal class="xl" :show="showManageApplicationsModal" :footer="false"
-               :title="group.name" @close="closeManageApplications">
-            <ManageApplicationsModal :group="group" />    
-        </Modal>
         <Modal class="xl" :show="showInviteUsersModal" :footer="false"
                :title="$t('invite-users-to', {group: group.name})" @close="closeInviteUsersModal">
             <InviteUsersModal :group="group" />
@@ -39,25 +75,27 @@
 
 <script setup lang="ts">
 import {onBeforeMount, ref, PropType} from 'vue';
-import ManageGroupModal from '../ManageGroupModal.vue';
-import ManageApplicationsModal from '../ManageApplicationsModal.vue';
-import InviteUsersModal from '../InviteUsersModal.vue';
-import Blocklist from '../Blocklist.vue';
-import {daysSince, parseDateTime} from '/js/services/dateService';
+import InviteUsersModal from '../modals/InviteUsersModal.vue';
+import Blocklist from '../modals/Blocklist.vue';
 import {useGroupStore} from '/js/store/groupStore';
-import {useRoute, useRouter} from 'vue-router';
+import {useRouter} from 'vue-router';
 import {useI18n} from 'vue-i18n';
-import {GroupPage} from 'resources/types/group';
+import {Application, Group, GroupPage} from 'resources/types/group';
+import Editable from '/js/components/global/Editable.vue';
 
 const groupStore = useGroupStore();
-const route = useRoute();
 const router = useRouter();
 const {t} = useI18n();
 
-onBeforeMount(async() => {
-    await groupStore.fetchGroup(parseInt(String(route.params.id)));
-    loading.value = false;
+onBeforeMount(() => {
+    if (props.group.require_application) loadApplications();
 });
+
+async function loadApplications() {
+    loading.value = true;
+    applications.value = await groupStore.fetchApplications(props.group.id);
+    loading.value = false;
+}
 
 const props = defineProps({
     group: {
@@ -65,6 +103,8 @@ const props = defineProps({
         required: true,
     },
 });
+
+const applications = ref<Array<Application> | null>(null);
 
 const loading = ref(true);
 
@@ -76,20 +116,44 @@ async function deleteGroup() {
     }
 }
 
-const showManageGroupModal = ref(false);
-function manageGroup() {
-    showManageGroupModal.value = true;
+/**
+ * Manage group settings
+ */
+function save(group: Group) {
+    group.id = props.group.id;
+    groupStore.updateGroup(group);
 }
-function closeManageGroup() {
-    showManageGroupModal.value = false;
+function togglePublic() {
+    const group = {} as Group;
+    group.is_public = !props.group.is_public;
+    group.id = props.group.id;
+    groupStore.updateGroup(group);
 }
 
-const showManageApplicationsModal = ref(false);
-function manageApplications() {
-    showManageApplicationsModal.value = true;
+function toggleRequireApplication() {
+    const group = {} as Group;
+    group.require_application = !props.group.require_application;
+    group.id = props.group.id;
+    groupStore.updateGroup(group);
+    if (group.require_application) loadApplications();
 }
-function closeManageApplications() {
-    showManageApplicationsModal.value = false;
+
+/**
+ * Manage applications
+ */
+async function rejectApplication(applicationId: number) {
+    await groupStore.rejectApplication(applicationId);
+    loadApplications();
+}
+
+async function acceptApplication(applicationId: number) {
+    await groupStore.acceptApplication(applicationId);
+    loadApplications();
+}
+
+async function suspendApplication(applicationId: number) {
+    await groupStore.suspendApplication(applicationId);
+    loadApplications();
 }
 
 const showInviteUsersModal = ref(false);
@@ -107,4 +171,5 @@ function showBlocklist() {
 function closeBlocklistModal() {
     showBlocklistModal.value = false;
 }
+
 </script>
